requirements.txt
click==8.0.1
colorama==0.4.4
dataclasses==0.6
dnspython==1.16.0
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gunicorn==20.1.0
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysqlclient==2.0.3
passlib==1.7.4
PyJWT==2.1.0
pymongo==3.11.4
PyMySQL==1.0.2
SQLAlchemy==1.4.17
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1
